backend/pyproject.toml
[tool.poetry]
name = "ai-governance-hub"
version = "0.1.0"
description = "AI Model Governance & Compliance Hub"
authors = ["Admin <admin@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlmodel = "^0.0.14"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
pydantic = "^2.6.0"
python-multipart = "^0.0.7"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
reportlab = "^4.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
httpx = "^0.26.0"
pytest-asyncio = "^0.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
